refinements to drag behavior
each InteractionHandler now contains an array of DragUpdateListeners, each of which now uses the scaled coordinate difference from the drag start to update an expression. this will allow, for example, the x and y coordinates to determine a level set, and update the parameter that's the level value, rather than the x and y coordinates themselves. it also allows the updating of multiple parameters from a single drag event.